requirements.txt
tensorflow>=2.13.0
pillow>=10.0.0
matplotlib>=3.7.0
pandas>=2.0.0
opencv-python>=4.8.0
progressbar2>=4.2.0
wandb>=0.15.0
black>=23.0.0
gdown>=4.7.0